# openapi.yaml
openapi: 3.0.3
info:
  title: API de Productos y Ordenes
  description: >
    API REST para dar soporte a la venta de artículos, tanto presencial como
    por pedidos web. Expone los recursos Productos y Ordenes bajo un estilo
    arquitectónico de microservicios.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Entorno local (Docker)

tags:
  - name: Productos
  - name: Ordenes

paths:

  /api/productos:
    get:
      tags:
        - Productos
      summary: Listar productos
      description: Devuelve un listado de todos los productos disponibles.
      operationId: listarProductos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
              example:
                - id: "65a1f0c2e4b0a1a2b3c4d5e6"
                  nombre: "Notebook Lenovo ThinkPad"
                  descripcion: "Notebook Intel i7 16GB RAM"
                  precio: 1250.50
                  stock: 15
                  imagenes:
                    - "https://cdn.example.com/img/notebook-1.jpg"
                    - "https://cdn.example.com/img/notebook-2.jpg"

    post:
      tags: [Productos]
      summary: Crear un nuevo producto
      operationId: crearProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '400':
          description: Error de validación del cuerpo de la solicitud (@Valid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidacion'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 400
                errores:
                  precio: "no debe ser nulo"

  /api/productos/{id}:
    parameters:
      - $ref: '#/components/parameters/ProductoId'

    get:
      tags: [Productos]
      summary: Obtener un producto por identificador
      operationId: obtenerProducto
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '404':
          description: El producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 404
                mensaje: "El producto con id 65a1f0c2e4b0a1a2b3c4d5e6 no existe"

    put:
      tags: [Productos]
      summary: Reemplazar todos los datos de un producto existente
      operationId: reemplazarProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '400':
          description: Error de validación del cuerpo de la solicitud (@Valid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidacion'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 400
                errores:
                  precio: "no debe ser nulo"
        '404':
          description: El producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 404
                mensaje: "El producto con id 65a1f0c2e4b0a1a2b3c4d5e6 no existe"

    patch:
      tags: [Productos]
      summary: Actualizar parcialmente un producto existente
      operationId: actualizarProductoParcial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoPatch'
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '400':
          description: Error de validación del cuerpo de la solicitud (@Valid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidacion'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 400
                errores:
                  precio: "no debe ser nulo"
        '404':
          description: El producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 404
                mensaje: "El producto con id 65a1f0c2e4b0a1a2b3c4d5e6 no existe"

  /api/ordenes:
    get:
      tags: [Ordenes]
      summary: Obtener el listado de órdenes
      operationId: listarOrdenes
      responses:
        '200':
          description: Lista de órdenes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Orden'
              example:
                - id: "65b2a1d3f5c0b2b3c4d5e6f7"
                  email: "cliente@email.com"
                  direccion: "Av. Italia 3333, Maldonado"
                  telefono: "+59899111222"
                  estado: "CREATED"
                  fechaCreacion: "2026-06-21T14:30:00-03:00"
                  items:
                    - productoId: "65a1f0c2e4b0a1a2b3c4d5e6"
                      cantidad: 2
                    - productoId: "65a1f0c2e4b0a1a2b3c4d5e7"
                      cantidad: 5

    post:
      tags: [Ordenes]
      summary: Crear una nueva orden de compra
      description: >
        Crea la orden en estado CREATED y publica un evento en RabbitMQ para
        que el servicio de procesamiento verifique stock de forma asíncrona
        y la marque como READY_TO_DELIVERY o NO_STOCK.
      operationId: crearOrden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdenInput'
      responses:
        '201':
          description: Orden creada exitosamente en estado CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'
        '400':
          description: Error de validación del cuerpo de la solicitud (@Valid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidacion'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 400
                errores:
                  email: "debe ser una dirección de correo con formato correcto"
        '409':
          description: >
            Alguno de los productoId indicados no existe. NOTA: la verificación
            de stock ya no ocurre acá — se movió al servicio de procesamiento
            asíncrono, que marca la orden como NO_STOCK si corresponde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 409
                mensaje: "Producto no encontrado con ID: 65a1f0c2e4b0a1a2b3c4d5e6"

  /api/ordenes/{id}:
    parameters:
      - $ref: '#/components/parameters/OrdenId'

    get:
      tags: [Ordenes]
      summary: Obtener una orden por identificador
      operationId: obtenerOrden
      responses:
        '200':
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'
        '404':
          description: La orden no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 404
                mensaje: "La orden con id 65b2a1d3f5c0b2b3c4d5e6f7 no existe"

  /api/ordenes/{id}/detalle:
    parameters:
      - $ref: '#/components/parameters/OrdenId'

    get:
      tags: [Ordenes]
      summary: Obtener una orden con el detalle completo de cada producto solicitado
      operationId: obtenerDetalleOrden
      responses:
        '200':
          description: Detalle de la orden encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenDetalle'
        '404':
          description: La orden no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              example:
                timestamp: "2026-08-30T19:30:35.407942"
                status: 404
                mensaje: "La orden con id 65b2a1d3f5c0b2b3c4d5e6f7 no existe"

  /api/ordenes/{id}/factura:
    parameters:
      - $ref: '#/components/parameters/OrdenId'

    get:
      tags: [Ordenes]
      summary: Obtener la factura generada para una orden
      description: >
        Solo existe si el servicio de procesamiento asíncrono ya marcó la
        orden como READY_TO_DELIVERY. Si la orden sigue en CREATED o terminó
        en NO_STOCK, no hay factura y este endpoint responde 404.
      operationId: obtenerFacturaDeOrden
      responses:
        '200':
          description: Factura encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factura'
        '404':
          description: >
            La orden no existe, o existe pero todavía no tiene una factura
            generada (está en CREATED o en NO_STOCK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSimple'
              examples:
                ordenNoExiste:
                  summary: La orden no existe
                  value:
                    timestamp: "2026-08-30T19:30:35.407942"
                    status: 404
                    mensaje: "Orden no encontrada con ID: 65b2a1d3f5c0b2b3c4d5e6f7"
                sinFactura:
                  summary: La orden existe pero no tiene factura todavía
                  value:
                    timestamp: "2026-08-30T19:30:35.407942"
                    status: 404
                    mensaje: "La orden con ID 65b2a1d3f5c0b2b3c4d5e6f7 todavía no tiene una factura generada"

components:

  parameters:
    ProductoId:
      name: id
      in: path
      required: true
      description: Identificador del producto (ObjectId de MongoDB)
      schema:
        type: string
        example: "65a1f0c2e4b0a1a2b3c4d5e6"

    OrdenId:
      name: id
      in: path
      required: true
      description: Identificador de la orden (ObjectId de MongoDB)
      schema:
        type: string
        example: "65b2a1d3f5c0b2b3c4d5e6f7"

  schemas:

    Producto:
      type: object
      properties:
        id:
          type: string
          description: Identificador generado por MongoDB (ObjectId)
          example: "65a1f0c2e4b0a1a2b3c4d5e6"
        nombre:
          type: string
          example: "Notebook Lenovo ThinkPad"
        descripcion:
          type: string
          example: "Notebook Intel i7 16GB RAM"
        precio:
          type: number
          format: double
          example: 1250.50
        stock:
          type: integer
          example: 15
        imagenes:
          type: array
          description: Lista de links a imágenes del producto
          items:
            type: string
            format: uri
          example:
            - "https://cdn.example.com/img/notebook-1.jpg"
            - "https://cdn.example.com/img/notebook-2.jpg"

    ProductoInput:
      type: object
      required:
        - nombre
        - descripcion
        - precio
        - stock
      properties:
        nombre:
          type: string
          example: "Notebook Lenovo ThinkPad"
        descripcion:
          type: string
          example: "Notebook Intel i7 16GB RAM"
        precio:
          type: number
          format: double
          example: 1250.50
        stock:
          type: integer
          example: 15
        imagenes:
          type: array
          items:
            type: string
            format: uri

    ProductoPatch:
      type: object
      description: Campos a modificar; se admite cualquier subconjunto de los atributos del producto
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        precio:
          type: number
          format: double
        stock:
          type: integer
        imagenes:
          type: array
          items:
            type: string
            format: uri

    OrdenItem:
      type: object
      required:
        - productoId
        - cantidad
      properties:
        productoId:
          type: string
          description: Identificador del producto (ObjectId de MongoDB)
          example: "65a1f0c2e4b0a1a2b3c4d5e6"
        cantidad:
          type: integer
          example: 2

    Orden:
      type: object
      properties:
        id:
          type: string
          description: Identificador generado por MongoDB (ObjectId)
          example: "65b2a1d3f5c0b2b3c4d5e6f7"
        email:
          type: string
          format: email
          example: "cliente@email.com"
        direccion:
          type: string
          example: "Av. Italia 3333, Maldonado"
        telefono:
          type: string
          example: "+59899111222"
        estado:
          type: string
          description: >
            CREATED: recién creada, pendiente de procesamiento asíncrono.
            READY_TO_DELIVERY: el servicio de procesamiento confirmó stock
            y descontó las cantidades. NO_STOCK: no había stock suficiente
            para alguno de los productos solicitados.
          enum: [CREATED, READY_TO_DELIVERY, NO_STOCK]
          example: "CREATED"
        fechaCreacion:
          type: string
          format: date-time
          example: "2026-06-21T14:30:00-03:00"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrdenItem'

    OrdenInput:
      type: object
      required:
        - email
        - direccion
        - telefono
        - items
      properties:
        email:
          type: string
          format: email
          example: "cliente@email.com"
        direccion:
          type: string
          example: "Av. Italia 3333, Maldonado"
        telefono:
          type: string
          example: "+59899111222"
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrdenItem'

    OrdenDetalleItem:
      type: object
      description: Detalle completo del producto solicitado junto a la cantidad pedida
      properties:
        producto:
          $ref: '#/components/schemas/Producto'
        cantidad:
          type: integer
          example: 2

    OrdenDetalle:
      type: object
      properties:
        id:
          type: string
          description: Identificador generado por MongoDB (ObjectId)
          example: "65b2a1d3f5c0b2b3c4d5e6f7"
        email:
          type: string
          format: email
          example: "cliente@email.com"
        direccion:
          type: string
          example: "Av. Italia 3333, Maldonado"
        telefono:
          type: string
          example: "+59899111222"
        estado:
          type: string
          enum: [CREATED, READY_TO_DELIVERY, NO_STOCK]
          example: "CREATED"
        fechaCreacion:
          type: string
          format: date-time
          example: "2026-06-21T14:30:00-03:00"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrdenDetalleItem'

    FacturaItem:
      type: object
      description: Un ítem facturado, con el precio unitario congelado al momento del procesamiento
      properties:
        productoId:
          type: string
          description: Identificador del producto (ObjectId de MongoDB)
          example: "65a1f0c2e4b0a1a2b3c4d5e6"
        cantidad:
          type: integer
          example: 2
        precioUnitario:
          type: number
          format: double
          description: Precio del producto en el momento en que se procesó la orden
          example: 1250.50
        subtotal:
          type: number
          format: double
          description: Calculado como cantidad × precioUnitario (no se persiste, se calcula al responder)
          example: 2501.00

    Factura:
      type: object
      properties:
        id:
          type: string
          description: Identificador generado por MongoDB (ObjectId)
          example: "65c3b2e4f6d1c3d4e5f6a7b8"
        ordenId:
          type: string
          description: Identificador de la orden facturada
          example: "65b2a1d3f5c0b2b3c4d5e6f7"
        items:
          type: array
          items:
            $ref: '#/components/schemas/FacturaItem'
        montoTotal:
          type: number
          format: double
          description: Suma de los subtotales de todos los items
          example: 2501.00

    ErrorSimple:
      type: object
      description: >
        Estructura devuelta por GlobalExceptionHandler cuando se lanza una
        ResponseStatusException (usada para 404 y 409).
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-30T19:30:35.407942"
        status:
          type: integer
          example: 404
        mensaje:
          type: string
          example: "El recurso solicitado no existe"

    ErrorValidacion:
      type: object
      description: >
        Estructura devuelta por GlobalExceptionHandler cuando falla la
        validación @Valid del cuerpo de la solicitud (siempre 400).
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-30T19:30:35.407942"
        status:
          type: integer
          example: 400
        errores:
          type: object
          description: Mapa de nombre de campo -> mensaje de error de validación
          additionalProperties:
            type: string
          example:
            precio: "no debe ser nulo"
            email: "debe ser una dirección de correo con formato correcto"